fix: PSR-05

+++ contracts/ProtocolReserve/ProtocolShareReserve.sol
@@ -164,12 +164,12 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
         //because prime relies on getUnreleasedFunds and its return value may change after config update
         _accrueAndReleaseFundsToPrime();
 
-        for (uint i = 0; i < configs.length; ++i) {
+        for (uint256 i = 0; i < configs.length; ++i) {
             DistributionConfig memory _config = configs[i];
             require(_config.destination != address(0), "ProtocolShareReserve: Destination address invalid");
 
             bool updated = false;
-            for (uint j = 0; j < distributionTargets.length; ++j) {
+            for (uint256 j = 0; j < distributionTargets.length; ++j) {
                 DistributionConfig storage config = distributionTargets[j];
 
                 if (_config.schema == config.schema && config.destination == _config.destination) {
@@ -201,7 +201,7 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
     function releaseFunds(address comptroller, address[] memory assets) external {
         _accruePrimeInterest();
 
-        for (uint i = 0; i < assets.length; ++i) {
+        for (uint256 i = 0; i < assets.length; ++i) {
             _releaseFund(comptroller, assets[i]);
         }
     }
@@ -219,7 +219,7 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
         address destination,
         address asset
     ) external view returns (uint256) {
-        for (uint i = 0; i < distributionTargets.length; ++i) {
+        for (uint256 i = 0; i < distributionTargets.length; ++i) {
             DistributionConfig storage _config = distributionTargets[i];
             if (_config.schema == schema && _config.destination == destination) {
                 uint256 total = assetsReserves[comptroller][asset][schema];
@@ -274,7 +274,7 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
      */
     function _accrueAndReleaseFundsToPrime() internal {
         address[] memory markets = IPrime(prime).allMarkets();
-        for (uint i = 0; i < markets.length; ++i) {
+        for (uint256 i = 0; i < markets.length; ++i) {
             address market = markets[i];
             IPrime(prime).accrueInterest(market);
             _releaseFund(CORE_POOL_COMPTROLLER, _getUnderlying(market));
@@ -287,7 +287,7 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
      */
     function _accruePrimeInterest() internal {
         address[] memory markets = IPrime(prime).allMarkets();
-        for (uint i = 0; i < markets.length; ++i) {
+        for (uint256 i = 0; i < markets.length; ++i) {
             address market = markets[i];
             IPrime(prime).accrueInterest(market);
         }
@@ -295,10 +295,10 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
 
     function _releaseFund(address comptroller, address asset) internal {
         uint256 totalSchemas = uint256(type(Schema).max) + 1;
-        uint[] memory schemaBalances = new uint[](totalSchemas);
+        uint256[] memory schemaBalances = new uint256[](totalSchemas);
         uint256 totalBalance;
 
-        for (uint schemaValue; schemaValue < totalSchemas; ++schemaValue) {
+        for (uint256 schemaValue; schemaValue < totalSchemas; ++schemaValue) {
             schemaBalances[schemaValue] = assetsReserves[comptroller][asset][Schema(schemaValue)];
             totalBalance += schemaBalances[schemaValue];
         }
@@ -307,9 +307,9 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
             return;
         }
 
-        uint[] memory totalTransferAmounts = new uint[](totalSchemas);
+        uint256[] memory totalTransferAmounts = new uint256[](totalSchemas);
 
-        for (uint i = 0; i < distributionTargets.length; ++i) {
+        for (uint256 i = 0; i < distributionTargets.length; ++i) {
             DistributionConfig memory _config = distributionTargets[i];
 
             uint256 transferAmount = (schemaBalances[uint256(_config.schema)] * _config.percentage) / MAX_PERCENT;
@@ -321,8 +321,8 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
             emit AssetReleased(_config.destination, asset, _config.schema, _config.percentage, transferAmount);
         }
 
-        uint[] memory newSchemaBalances = new uint[](totalSchemas);
-        for (uint schemaValue = 0; schemaValue < totalSchemas; ++schemaValue) {
+        uint256[] memory newSchemaBalances = new uint256[](totalSchemas);
+        for (uint256 schemaValue = 0; schemaValue < totalSchemas; ++schemaValue) {
             newSchemaBalances[schemaValue] = schemaBalances[schemaValue] - totalTransferAmounts[schemaValue];
             assetsReserves[comptroller][asset][Schema(schemaValue)] = newSchemaBalances[schemaValue];
             totalAssetReserve[asset] = totalAssetReserve[asset] - totalTransferAmounts[schemaValue];
@@ -352,14 +352,14 @@ contract ProtocolShareReserve is AccessControlledV8, IProtocolShareReserve {
 
     function _ensurePercentages() internal view {
         uint256 totalSchemas = uint256(type(Schema).max) + 1;
-        uint[] memory totalPercentages = new uint[](totalSchemas);
+        uint256[] memory totalPercentages = new uint256[](totalSchemas);
 
-        for (uint i = 0; i < distributionTargets.length; ++i) {
+        for (uint256 i = 0; i < distributionTargets.length; ++i) {
             DistributionConfig memory config = distributionTargets[i];
             totalPercentages[uint256(config.schema)] += config.percentage;
         }
 
-        for (uint schemaValue = 0; schemaValue <= totalSchemas - 1; ++schemaValue) {
+        for (uint256 schemaValue = 0; schemaValue <= totalSchemas - 1; ++schemaValue) {
             if (totalPercentages[schemaValue] != MAX_PERCENT && totalPercentages[schemaValue] != 0)
                 revert InvalidTotalPercentage();
         }
